Hyojin Bae Week2 Algorithm Study
